+++ config/initializers/devise.rb
@@ -4,7 +4,7 @@ Devise.setup do |config|
   # The secret key used by Devise. Devise uses this key to generate
   # random tokens. Changing this key will render invalid all existing
   # confirmation, reset password and unlock tokens in the database.
-  config.secret_key = APP_CONFIG['devise_secret']
+  config.secret_key = ENV['devise_secret']
   
   # ==> Mailer Configuration
   # Configure the e-mail address which will be shown in Devise::Mailer,