+++ db/migrate/20150227113121_change_api_user_default_value.rb
@@ -1,0 +1,5 @@
+class ChangeApiUserDefaultValue < ActiveRecord::Migration
+  def change
+    change_column_default :users, :active, nil
+  end
+end